Perfect - Make original buffer use a disk backed mapping to save memory. - rearrange api function headers for search etc. - fix bugs with iterators. - support uint64_t for all byte and line offsets to allow for much larger files to be opened. - and more minor bug fixes.

+++ include/vase/constants.h
@@ -2,8 +2,8 @@
 
 #include "pch.h"
 
-constexpr uint32_t APPEND_CHUNK_SIZE = 64 * 1024;
-constexpr uint32_t PETAL_SIZE_MAX = 32 * 1024;
+constexpr uint64_t APPEND_CHUNK_SIZE = 64 * 1024;
+constexpr uint64_t PETAL_SIZE_MAX = 32 * 1024;
 
 static_assert(
   APPEND_CHUNK_SIZE > PETAL_SIZE_MAX,
